Move str2path() from utils.api to utils.file
