+++ src/uwtools/tests/utils/test_file.py
@@ -112,6 +112,18 @@ def test_resource_path():
     assert file.resource_path().is_dir()
 
 
+@mark.parametrize("val", [Path("/some/path"), {"foo": 88}])
+def test_str2path_passthrough(val):
+    assert file.str2path(val) == val
+
+
+def test_str2path_convert():
+    val = "/some/path"
+    result = file.str2path(val)
+    assert isinstance(result, Path)
+    assert result == Path(val)
+
+
 def test_writable_file(tmp_path):
     apath = tmp_path / "afile"
     with file.writable(filepath=apath) as f:
