+++ src/uwtools/utils/file.py
@@ -102,6 +102,15 @@ def resource_path(suffix: str = "") -> Path:
         return prefix / suffix
 
 
+def str2path(val: Any) -> Any:
+    """
+    Return str value as Path, other types unmodified.
+
+    :param val: Any value.
+    """
+    return Path(val) if isinstance(val, str) else val
+
+
 @contextmanager
 def writable(filepath: Optional[Path] = None, mode: str = "w") -> Generator[IO, None, None]:
     """
